requirements.txt
pandas==2.18.4
sklearn==1.3.4
fastapi-cli==0.0.7
mlflow==2.20.2